Move rate limit code to `ratelimit` package.
The rate limiting code previously lived in the `cmd` package without
a clear justification for why. This commit moves the rate limiting code
to its own `ratelimit` package and updates import paths as required.
Notably all references from the `cmd` package's exported `LoadRateLimitPolicies`,
`RateLimitPolicy`, and `RateLimitConfig` were moved to use `ratelimit`.
This removed the `cmd` import from a couple of callers (nice!).
